+++ examples/index.example.ts
@@ -1,3 +1,10 @@
 import { createLogger } from "../src";
 
-export const logger = await createLogger({path: 'application-logs'});
+export const logger = await createLogger({
+    file: {
+        path: 'example-logs',
+    },
+    logger: {
+        service: 'example-service',
+    }
+});