+++ tests/createLogger.test.ts
@@ -2,8 +2,15 @@ import { createLogger } from "../src/index.js"
 import { Logger } from "../src/logger.js"
 import { describe, it, expect } from "vitest"
 
-const path = 'application-logs';
-const logger = await createLogger({ path });
+const path = 'tests-logs'
+const logger = await createLogger({
+    file: {
+        path
+    },
+    logger: {
+        service: 'test',
+    }
+});
 
 describe('createLogger', () => {